Disable toolchain libgcc checks if they are not needed
This disables the checks of the toolchain libgcc that serve no purpose
(or even cause false positives if the toolchain does not provide a
libgcc at all) in case we do not need the toolchain libgcc (i.e. the
CONFIG_BID_USE_TOOLCHAIN_LIBGCC option is disabled).

If CONFIG_BID_USE_TOOLCHAIN_LIBGCC is enabled, there is no change in
functionality.

Change-Id: If25f2e046f9635681db58e9c382d6b0f5dbeccd1